Extract risk summary response module
